make authenticate copy r->user from r->main in subrequests
If we are in a subrequest, r->main is defined, r->prev is not.
authenticate will now copy r->user from r->main in this instance.  Since
rubrequests do not come from the client directly, the intial user would
have been authenticated in the inital or main request.  This avoids
having to re-check authentication in subrequests.

This also fixes DirectoryIndexes on auth required directories under
apache 2.4.

This fixes #5