remove CGI.pm dependency and replace with smaller/lighter modules
CGI.pm has been removed from perl core, which was the primary reason we
used it in the first place.  Its a bloated mess, and full of quirks.  It
has been replaced here with three smaller modules that are much simpler:
        - HTTP::Body
        - WWW::Form::UrlEncoded
        - Hash::MultiValue

A Runtime Recommends dependency is declared for
WWW::Form::UrlEncoded::XS which will automatically be used if present.